pyproject.toml
[project]
name = "story-factory-ai"
version = "0.1.0"
description = "Generador agéntico de cuentos con control de calidad y persistencia"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langgraph", # El motor del flujo
    "langchain-openai", # Integración con modelos de OpenAI
    "langchain-core", # Abstracciones base
    "python-dotenv", # Gestión de variables de entorno (.env)
    "pydantic", # Validación de datos (crucial para producción)
    "langgraph-checkpoint-sqlite>=3.0.3",
]

[tool.uv]
dev-dependencies = [
    "pytest",              # Para nuestros tests unitarios
    "langsmith",           # Para observabilidad y debug en producción
]
